requirements.txt
numpy>=1.15.0
pandas>=0.25.1
matplotlib>=3.0.0
opencv-python>=3.4.5.20
tensorflow-gpu>=2.1.0
